image-bot/requirements.txt
playwright>=1.42.0
playwright-stealth>=1.0.6
gspread>=6.0.0
google-api-python-client>=2.122.0
oauth2client>=4.1.3
python-dotenv>=1.0.1

opencv-python-headless
numpy
Pillow
rembg>=2.0.50
